requirements.txt
appnope==0.1.0
astroid==1.5.3
atomicwrites==1.1.5
attrs==18.1.0
autopep8==1.3.4
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.48.0
boto3==1.7.40
botocore==1.10.40
bz2file==0.98
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
cycler==0.10.0
decorator==4.3.0
dlib==19.13.1
docutils==0.14
enum34==1.1.6
face-recognition==1.2.2
face-recognition-models==0.3.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imageio==2.3.0
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.2.15
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
Keras==2.1.2
lazy==1.3
lazy-object-proxy==1.3.1
Markdown==2.6.10
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.2.0
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==4.3.0
pluggy==0.6.0
powerline-status==2.6
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
py==1.5.3
pycodestyle==2.3.1
Pygments==2.2.0
pylint==1.7.5
pyparsing==2.2.0
pytest==3.6.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
QueryableList==3.1.0
requests==2.18.4
rope==0.10.7
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.0
selenium==3.11.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.7
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
tinydb==3.8.1.post1
tqdm==4.23.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
xvfbwrapper==0.2.9
